moved abstraction parts to a separate project
