+++ Business/Business.csproj
@@ -7,12 +7,13 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Data\Data.csproj"/>
-        <ProjectReference Include="..\Cqrs\Cqrs.csproj"/>
+        <ProjectReference Include="..\Data\Data.csproj" />
+        <ProjectReference Include="..\Cqrs\Cqrs.csproj" />
+        <ProjectReference Include="..\Mapper\Mapper.csproj" />
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="automapper" Version="12.0.1"/>
+        <PackageReference Include="automapper" Version="12.0.1" />
     </ItemGroup>
     
 </Project>
