+++ Business/Mapper/Extensions/ServiceCollectionExtensions.cs
@@ -1,0 +1,13 @@
+
+using Mapper;
+using Mapper.Extensions;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Business.Mapper.Extensions;
+
+internal static class ServiceCollectionExtensions
+{
+    public static IServiceCollection AddMapper(this IServiceCollection source)
+        => source.AddMapper<DefaultMapper>(typeof(ServiceCollectionExtensions));
+
+}
