+++ DisplayPartsInventory.sln
@@ -11,6 +11,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Cqrs", "Cqrs\Cqrs.csproj", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestingConsoleApp", "TestingConsoleApp\TestingConsoleApp.csproj", "{447703E4-14F7-432A-B0DF-59CD0C540734}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mapper", "Mapper\Mapper.csproj", "{2548A455-D9F1-49C8-ACDE-19220F1B1B71}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -36,5 +38,9 @@ Global
 		{447703E4-14F7-432A-B0DF-59CD0C540734}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{447703E4-14F7-432A-B0DF-59CD0C540734}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{447703E4-14F7-432A-B0DF-59CD0C540734}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2548A455-D9F1-49C8-ACDE-19220F1B1B71}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2548A455-D9F1-49C8-ACDE-19220F1B1B71}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2548A455-D9F1-49C8-ACDE-19220F1B1B71}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2548A455-D9F1-49C8-ACDE-19220F1B1B71}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
