+++ TestingConsoleApp/Program.cs
@@ -1,5 +1,4 @@
 ﻿using AutoMapper;
-using Business.Mapper.Abstractions;
 using Data.Dtos;
 using Data.Entities;
 using Microsoft.Extensions.DependencyInjection;
